refactor: replace vue-slick-carousel with swiper
