+++ components/st-home-carousel.vue
@@ -1,78 +1,33 @@
 <template>
   <div class="st-home-carousel">
-    <!-- TODO: USE NEW CAROUSEL -->
-    <!-- <vue-slick-carousel v-if="items.length > 0" v-bind="settings"> -->
-    <nuxt-link
-      v-for="(item, index) in items"
-      :key="`carousel-item-${index}`"
-      :to="item.href"
-      class="st-home-carousel__item"
-    >
-      <img
-        class="st-home-carousel__image"
-        :alt="item.image.alt"
-        :src="imageFallbackSrc(item.image.directusAssetId)"
-        :srcset="imageSrcSet(item.image.directusAssetId)"
-        :sizes="imgTagSizes"
-      />
-      <div class="st-home-carousel__title-fade"></div>
-      <h3 class="st-home-carousel__title">
-        {{ item.caption }}
-      </h3>
-    </nuxt-link>
-    <!-- </vue-slick-carousel> -->
+    <client-only>
+      <swiper-container
+        class="st-home-carousel__swiper"
+        css-mode="true"
+        :pagination="{ clickable: true }"
+        :autoplay="{ delay: 5000, pauseOnMouseEnter: true }"
+        rewind="true"
+      >
+        <swiper-slide v-for="(item, index) in items" :key="`carousel-item-${index}`">
+          <st-home-carousel-item :item="item" />
+        </swiper-slide>
+      </swiper-container>
+      <template #fallback>
+        <st-home-carousel-item :item="items[0]" class="st-home-carousel__item-fallback" />
+      </template>
+    </client-only>
   </div>
 </template>
 
 <script setup lang="ts">
-import { getAssetSrcSet, getAssetURL } from '~/plugins/06.directus';
+import type { CarouselItem } from './st-home-carousel-item.vue';
 
-const runtimeConfig = useRuntimeConfig();
-const appConfig = useAppConfig();
-
-export interface CarouselItem {
-  image: {
-    directusAssetId: string;
-    alt?: string;
-  };
-  caption: string;
-  href: string;
-}
-
-// TODO: Switch to Swiper when we'll be using Vue 3.
-//       https://swiperjs.com/vue
 defineProps({
   items: {
     type: Array as PropType<CarouselItem[]>,
     required: true,
   },
 });
-// const settings = ref({
-//   autoplay: true,
-//   autoplaySpeed: 5000,
-//   slidesToShow: 1,
-//   dots: true,
-//   arrows: false,
-// });
-const imgTagSizes = ref<string>('');
-
-onMounted(() => {
-  const bodyStyles = window.getComputedStyle(document.body);
-  const lXlBreakpoint = bodyStyles.getPropertyValue('--st-breakpoint-l-xl');
-  imgTagSizes.value = `(min-width: ${lXlBreakpoint}) ${lXlBreakpoint}, 100vw`;
-});
-
-const imageFallbackSrc = (assetId: string): string => {
-  return getAssetURL(runtimeConfig.public.cmsURL, assetId, {
-    width: appConfig.keyVisualSizes[0],
-  });
-};
-
-const imageSrcSet = (assetId: string): string => {
-  return getAssetSrcSet(runtimeConfig.public.cmsURL, assetId, {
-    widths: appConfig.keyVisualSizes,
-  });
-};
 </script>
 
 <style scoped>
@@ -83,105 +38,36 @@ const imageSrcSet = (assetId: string): string => {
   margin-left: calc(-1 * var(--st-length-main-content-side-padding));
 }
 
-.st-home-carousel__image {
-  display: block;
-  width: 100%;
-  object-fit: cover;
-  background-image: linear-gradient(gray 100%, transparent 0);
-  aspect-ratio: 3 / 2;
+.st-home-carousel__swiper::part(pagination) {
+  position: static;
+  margin-top: var(--st-length-spacing-xs);
 }
 
-.st-home-carousel__item {
-  position: relative;
-  display: block;
+.st-home-carousel__swiper::part(bullet) {
+  background-color: var(--st-color-home-carousel-dot);
+  opacity: 1;
+  width: 10px;
+  height: 10px;
+  margin: 0 calc(var(--st-length-spacing-xs) / 2);
+  transition: background-color 0.3s ease;
 }
 
-.st-home-carousel__title {
-  position: absolute;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  padding: 0 var(--st-length-spacing-xs);
+.st-home-carousel__swiper::part(bullet-active) {
+  background-color: var(--st-color-home-carousel-dot-active);
+  width: 12px;
+  height: 12px;
+  margin: 0 calc(var(--st-length-spacing-xs) / 2);
+  transition: background-color 0.3s ease;
+}
+
+.st-home-carousel__item-fallback {
   margin-bottom: var(--st-length-spacing-xs);
-  color: var(--st-color-home-carousel-title-foreground);
-  font-weight: 900;
-  font-size: 2em;
-
-  /* PostCSS doesn't support adding prefix for line-clamp yet: https://github.com/postcss/autoprefixer/issues/1322 */
-  /* stylelint-disable-next-line value-no-vendor-prefix */
-  display: -webkit-box;
-  -webkit-line-clamp: 2; /* That's weird, but the -webkit- prefix works also for Firefox */
-  -webkit-box-orient: vertical;
-  line-clamp: 2;
-  overflow: hidden;
-  z-index: 0;
-}
-
-.st-home-carousel__title-fade {
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  height: 9rem;
-  background: linear-gradient(to bottom, rgba(0 0 0 / 0%), rgba(0 0 0 / 60%));
-}
-
-@media (--sm-and-up) {
-  .st-home-carousel__title {
-    font-size: 2.25em;
-  }
-
-  .st-home-carousel__title-fade {
-    height: 10rem;
-  }
-}
-
-@media (--md-and-up) {
-  .st-home-carousel__title {
-    padding: var(--st-length-spacing-xs) var(--st-length-spacing-s);
-    font-size: 2.5em;
-  }
-
-  .st-home-carousel__title-fade {
-    height: 11rem;
-  }
-}
-
-@media (--lg-and-up) {
-  .st-home-carousel__title {
-    padding: var(--st-length-spacing-xs) var(--st-length-spacing-m);
-    font-size: 3em;
-  }
-
-  .st-home-carousel__title-fade {
-    height: 12rem;
-  }
 }
 
 @media (--xl-and-up) {
   .st-home-carousel {
     width: var(--st-breakpoint-l-xl);
     margin-left: calc(-1 * (var(--st-breakpoint-l-xl) - var(--st-length-main-content-max-width)) / 2);
-  }
-}
-
-@media (orientation: landscape) {
-  .st-home-carousel__image {
-    aspect-ratio: 2 / 1;
-  }
-
-  @supports not (aspect-ratio: 2 / 1) {
-    .st-home-carousel__image::before {
-      float: left;
-      padding-top: 40%;
-      content: '';
-    }
-
-    .st-home-carousel__image::after {
-      display: block;
-      content: '';
-      clear: both;
-    }
   }
 }
 </style>
