+++ package.json
@@ -41,6 +41,7 @@
     "pinia": "^2.1.7",
     "postcss-custom-media": "^10.0.6",
     "slugify": "^1.6.6",
+    "swiper": "^11.1.4",
     "vue": "^3.4.27",
     "vue-router": "^4.3.2"
   },
