+++ plugins/swiper.ts
@@ -1,0 +1,6 @@
+import { register } from 'swiper/element/bundle';
+import 'swiper/css/bundle';
+
+export default defineNuxtPlugin(() => {
+  register();
+});
